+++ test/support/test_config.ex
@@ -1,0 +1,14 @@
+defmodule Ravenx.Test.TestConfig do
+
+  def test_multiple(_) do
+    %{
+      token2: "MySecondSecretToken"
+    }
+  end
+
+  def test_module(_) do
+    %{
+      token2: "MySecondSecretToken"
+    }
+  end
+end